Add a catagory table and its associations. Also add data to the database

+++ db/migrate/20220314123549_create_foods.rb
@@ -3,7 +3,7 @@ class CreateFoods < ActiveRecord::Migration[7.0]
     create_table :foods do |t|
       t.string :name
       t.string :unit
-      t.integer :price
+      t.numeric :price
 
       t.timestamps
     end
